Fix a bug in hm2 encoder velocity estimation
When the encoder changes direction by a single edge, we can not reliably
estimate velocity.  The encoder might be "balancing on an edge", leading
to potentially very small delta-T between edges, leading to misleadingly
large velocity estimation.  In this situation it's better to report
vel=0 and wait for more data to come in.