Don't change hm2 stepgen position when .enable is low
This commit makes the hm2 stepgen *not* change its .counts and
.position-fb pins when .enable is low.

When emc2 disables a stepgen (for example when the user hits E-stop,
or when the joint ferrors), it copies the joint's position-fb to its
position-cmd, to avoid abrupt motion when the joint is enabled again.
Therefore my previous commit, which made the stepgen count and position-fb
track the position-cmd from emc2, was redundant and potentially risky.

Thanks to Jeff Epler for setting me straight on this.