remove the regmap, point to the one in the firmware repo
For a while we had a separate fork of the hostmot2 regmap file in the
emc2 git repo.  It drifted out of sync with the canonical one, which
is kept in the hostmot2-firmware repo.  This commit removes the emc2
copy of the regmap and replaces it with a Readme file that points to
the canonical version (via gitweb).