fix the PD Secondary Pin mask
The Secondary Pin field of the Pin Descriptor uses bit 7 (the most
significant bit) to mean "this pin is an output".  It does not use bit
6 for anything, that bit is available to indicate the actual use of the
pin, just like bits 0-5.

This commit changes the mask applied to the Secondary Pin field so it
just masks out bit 7, not bit 7 and 6.