fix a bug in stepgen mode setting
This commit fixes a bug that caused the stepgen mode to usually be written
out to the FPGA each servo cycle.  In some extremely rare conditions, it
could cause the stepgen mode to *not* be written out when it was needed.

It now writes the stepgen mode out whenever it's changed, and doesn't
write it out if it's not changed.