remove obsolete whiny documentation
Back when the hostmot2 firmware was part of the LinuxCNC repo, the user
needed to pay special attention for firmware loading to work right in
RIP builds.

Also, the way the hostmot2 driver does firmware loading is very different
from how the predecessors (m5i20 and the old stand-alone 7i43 hm2 driver)
did it.

So I wrote a document explaining my reasoning for changing how firmware
loading was handled.  This document is no longer helpful, so i'm
removing it.