Add a pin width limit to the Hostmot2 Stepgen
The recent addition of support for table mode to the Hostmot2 stepgen had an unfortunate
and unintended consequence. Some of the existing bitfiles did have wide stepgens included
but this was masked by the driver. Users that have configured these stepgen pins as GPIO
are likely to find that important parts of their machines no longer work.
This patch adds an extra parameter to the hm2 modparams so that system builders
have to actively choose to use the wide stepgens.

Signed-off-by: andy pugh <andy@bodgesoc.org>