hostmot2: bspi: Use proper function pointers
This is a potentially incompatible change, because formerly it was OK
to pass incompatible function pointer types and sometimes this would
"just work" (for instance, when they differed only by the type of the
pointer argument, or when the return type was erroneously void instead
of int), but the fix is easy (use correct function prototype;
cast argument to local pointer type; return 0 in the case of success)
and the change to mesa_7i65.comp in this commit provides an example of
such a fix.

This is preparatory work to fixing a diagnostic in bspi.c that the return
value of the read and write functions was ignored.