hm2-eth: system() is harmful to realtime priority
Both PCW and I saw that after the addition of calls to system(), there
were frequently 'realtime delays' logged at startup, particularly when
running with a 2kHz servo thread.  It does not appear to be the fault
of the command itself being executed, because running 'iptables -L fnord'
in another window while realtime is running does not cause realtime delays.

On Linux, based on looking at strace output, it appears that system() is
implemented in terms of clone(fork)+exec, while posix_spawn() is
implemented in terms of vfork+exec.  It's not clear why this difference
preserves realtime performance (at least for me), but if it works we'll
take it.