hm2: Fix format string security warning
The idiom
    printf(not_a_literal)
is warned about, because in the case that not_a_literal is under an
adversary's control, it could be made to contain %-substitutions in a
way that is a security flaw.  The correct replacement is
    printf("%s", not_a_literal)

because rtapi_dev_set_name is declared as a printf-style function, it
gets the same warning.  While it may be the case that some analysis
would show the string was never under the adversary's control, it's best
to rework the code so it's not a concern.