hm2: fix uninitialized variable warning in ioport
Before this commit, if the "pin" argument passed into count_instances()
was not found (because it was greater than the number of pins available,
or less than 0), then the "this_instance" argument would never get set,
and the calling code (in hm2_ioport_gpio_export_hal()) would use an
uninitialized value.

This commit fixes that problem.

This commit also renames a bunch of internal variables for clarity.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>