hm2_eth: detect late read packets
There are 16 bytes of scratch area in memory space 4
("timer/utility").  Write the number of this read packet there and
then read it back, in order to verify that the received packet is
the right one.  In case of realtime delays, we may appear to have a
missing packet at cycle N, but by cycle N+1 the missing read
response packet N has arrived.  In this case, we need to see that we
received the wrong packet (by wrong confirm_read_cnt value) and look
back to hopefully receive the right packet for cycle N+1.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>