hm2_7i34. hm2_7i90: Don't silently fail with blank config strings
Previously the number of config strings was used to hint to the hm2_epp
drivers how many cards to look for (and two default ioaddresses were supplied
If no config strings were found then no cards were looked for and the
driver silently quit with no error.
With the 7i90 card, which uses an eeprom, a blank config string is a likely
default, and the config modparam is not mentioned in the manpage.
This patch uses the number of ioaddresses supplied as the count hint.
For backwards compatibility it supplies the original defaults explicitly if
none are supplied and suppresses (some) error messages.
This workaround is likely to be removed in 2.8.

Signed-off-by: andypugh <andy@bodgesoc.org>