hm2: allow the llio to override the "connector-pin" name of IOs
Before this commit, the llio driver would specify the names of its IO
connectors (eg, "P1", "TB2", etc), the IDROM would specify the number
of pins per connector (eg, 24 for the standard 50-pin connectors, 17
for DB-25 connectors), and the "pins" code in the hostmot2 driver would
assign IOs sequentially to the pins on the connectors.

That doesn't work for the 7i96, and potentially for other Any-IO boards
where the mapping from IO to connector/pin is not straight-forward.

This commit introduces a new optional field that the llio driver can
provide when registering a board with the hostmot2 framework: an array
of char*, indexed by IO number, providing the string that should be
reported to the user for the connector/pin mapping of that IO.