BUGFIX: hostmot2 encoder quadrature error reporting bug. Previous to this patch the quadrature errors were not reported correctly and could be lost.
This patch requires a firmware change in the the encoder firmware.
Writing 1 to the quadrature error bit in the encoder control register will reset the error and writing 0 to it will do nothing.
Previous driver had a bug in error handling, where it did not report the quadrature error properly.
Using this patch with unpatched firmware is harmless as the encoder quadrature error did not work to start with.