hm2_pktuart: Support changes in Rx and Tx modules V2
Higher-resolution timers for better low-frequency performance and better
robustness to popping data which is not there.

Adds two new setup functions and a dedicated reset function.

The original hm2_pktuart_setup is left for compatibility but is now deprecated.
(It will work with the V2 uarts but can't set the filter for slow baudrates)




Signed-off-by: andypugh <andy@bodgesoc.org>